requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.1.14
distlib==0.3.6
Django==4.2
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
filelock==3.11.0
idna==3.4
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2==2.9.6
pypiwin32==223
pypng==0.20220715.0
pyTelegramBotAPI==4.10.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
qrcode==7.4.2
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
